Add support for hook command
Hook command is executed at the pre-update and post-update stages.
In contrast to the existing postupdatecmd option, the hook command
in the post-update stage is called (with different arguments) both
when the update succeeds and when an error occurs. Also, the hook
command is always called without additional IPC calls.

The hook command can be set via the command line option '-X, --hook'
or via the swupdate.cfg configuration file in the 'globals' section
with option 'hookcmd'.

The functionality of existing 'preupdatecmd' and 'postupdatecmd'
options remains unchanged.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>